pyproject.toml
[project]
name = "ecourt-gather"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex",email = "ink404dot@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pyodbc (>=5.3.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
